requirements.txt
tqdm
vietocr==0.3.11
fastai==1.0.60
LMDB
Pillow
opencv-python
tensorboardX
editdistance
synthtiger
numpy
opencv-python
addict
six
pillow>=6.0,<7.0
lmdb
nltk
terminaltables
albumentations

